LayoutTests/fast/forms/form-control-refresh/button-like-controls-text-color-matches.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<input id="select" type="select">
<input id="input_submit" type="submit">
<input id="input_button" type="button">
<button id="submit" type="submit"></button>
<button id="button"></button>
<script>
    description("This tests that buttons, submit buttons, and select buttons all have the same CSS color.");
    shouldBeTrue("select.style.color == input_submit.style.color");
    shouldBeTrue("select.style.color == input_button.style.color");
    shouldBeTrue("select.style.color == submit.style.color");
    shouldBeTrue("select.style.color == button.style.color");
</script>
</body>
</html>
